Copy constructor of User should now be the correct solution.
Although logically unique ptr is the correct type for storing Posts, we end up encountering an issue due to the generic nature of the implementation of LinkedBag.

We would change this, but since we are not allowed to, I will stick to the implementation that uses shared ptr.

==========

BUGS:

Post does not polymorphically copy. Implement a "clone" method that returns a shared ptr to a newly allocated copy of the post.